+++ core/pioreactor/automations/dosing/pid_morbidostat.py
@@ -5,9 +5,10 @@ from typing import Any
 from pioreactor import structs
 from pioreactor.automations import events
 from pioreactor.automations.dosing.base import DosingAutomationJob
+from pioreactor.background_jobs.dosing_automation import (
+    check_pump_calibrations_and_pwm_channels_are_configured,
+)
 from pioreactor.config import config
-from pioreactor.exc import CalibrationError
-from pioreactor.utils import local_persistent_storage
 from pioreactor.utils.streaming_calculations import PID
 
 
@@ -32,17 +33,13 @@ class PIDMorbidostat(DosingAutomationJob):
         skip_first_run: bool | str | int = False,
         **kwargs: Any,
     ) -> None:
-        super(PIDMorbidostat, self).__init__(**kwargs)
         assert target_normalized_od is not None, "`target_normalized_od` must be set"
         assert target_growth_rate is not None, "`target_growth_rate` must be set"
 
-        with local_persistent_storage("active_calibrations") as cache:
-            if "media_pump" not in cache:
-                raise CalibrationError("Media pump calibration must be performed first.")
-            elif "waste_pump" not in cache:
-                raise CalibrationError("Waste pump calibration must be performed first.")
-            elif "alt_media_pump" not in cache:
-                raise CalibrationError("Alt-Media pump calibration must be performed first.")
+        check_pump_calibrations_and_pwm_channels_are_configured(
+            ("media_pump", "waste_pump", "alt_media_pump")
+        )
+        super(PIDMorbidostat, self).__init__(**kwargs)
 
         self.set_target_growth_rate(target_growth_rate)
         self.target_normalized_od = float(target_normalized_od)